Synthesizer method renaming + new methods, update SynthConfig params (ev reading) (#63)
* refactor: `Synth::handle_event` -> `Synth::push_event`

* feat: `Synth::push_timed_event`
The ability to push a single event played sometime in the future

* refactor: SongInfo -> LoopedSong

Created store on res for playing and pausing

* refactor: Synth::play_audio now requires mutable reference

* feat: Synth::play, Synth::drop

* refactor: SongWriter::divide_events_into_channel

this method will return any hashmap where the hasher is BuildHasher and Default

* refactor: update the way synth event reading behaves

Essentially, the user can now handle a SynthEventReader if they choose. Alternatively,

They can utilize an event reader that's made in this crate if they don't wanna.

Updated the event type to include an Instant

* fix: synthreceiver is now spawned during startup

* chore: fix doc-test